Add template for TensorFlow library functions
This is mostly the same as the JAX template, with a few TensorFlow-specific modifications.

I've put in a few arbitrary axis reverals to match the current set of TensorFlow stubs. I'll remove them once we've confirmed that the stubs this template generates are identical to the current stubs.

Strictly speaking, this belongs to the same 'package' of changes as the next 3 CLs, so it really should all be in one CL, but that would be waaaaay too much to review in one go. Still, it means I'm taking some liberties with CL ordering here; one or two parts of this file correspond to later CLs. I've added comments indicating which parts these are.

PiperOrigin-RevId: 353233904